\name{Anscomb}
\alias{Anscomb}
\docType{data}
\title{
Anscomb dataset
}
\description{
Four pairs of data points for illustration of regression fallacies. (x1,y1) (x2,y2) (x3,y3) (x4 y4)
}
\usage{data(Anscomb)}
\format{
  A data frame with 11 observations on the following 8 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{y1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{y2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{y3}}{a numeric vector}
    \item{\code{x4}}{a numeric vector}
    \item{\code{y4}}{a numeric vector}
  }
}
\source{
Anscombe,A., Graphs in Statistical Analysis, American Statistician (circa 1974)
}
\examples{
data(Anscomb)
}
\keyword{datasets}
