\name{beerprod}
\Rdversion{1.1}
\alias{beerprod}
\docType{data}
\title{
BEERPROD dataset
}
\description{
Time series of monthly beer production in US in millions of barrels.
}
\usage{data(beerprod)}
\format{
  A data frame with 72 observations on the following variable.
  \describe{
    \item{\code{b_prod}}{a numeric vector with the amount of beer produced}
  }
}
\examples{
data(beerprod)
}
\keyword{datasets}
