\name{citizensol}
\alias{citizensol}
\docType{data}
\title{
CITIZENSOL dataset
}
\description{
Data from Citizen Soldiers by S. Ambrose (1997).
WWII US ARMY divisions
}
\usage{data(citizensol)}
\format{
  A data frame with 43 observations on the following 3 variables.
  \describe{
    \item{\code{days}}{days in combat on front line}
    \item{\code{ncasualties}}{total casualties}
    \item{\code{turnover}}{turnover in unit (e.g. 150\% means the total no. of 
    soldiers that served = 1.5 division strength)}
  }
}
\examples{
data(citizensol)
}
\keyword{datasets}
