\name{conret}
\alias{conret}
\docType{data}
\title{
CONRET dataset
}
\description{
Time series of returns on stock price indices for several countries.
}
\usage{data(conret)}
\format{
  A data frame with 107 observations on the following 23 variables.
  \describe{
    \item{\code{date}}{a numeric vector}
    \item{\code{australia}}{a numeric vector}
    \item{\code{austria}}{a numeric vector}
    \item{\code{belgium}}{a numeric vector}
    \item{\code{canada}}{a numeric vector}
    \item{\code{denmark}}{a numeric vector}
    \item{\code{finalnd}}{a numeric vector}
    \item{\code{france}}{a numeric vector}
    \item{\code{germany}}{a numeric vector}
    \item{\code{honkong}}{a numeric vector}
    \item{\code{irleland}}{a numeric vector}
    \item{\code{italy}}{a numeric vector}
    \item{\code{japan}}{a numeric vector}
    \item{\code{malaysia}}{a numeric vector}
    \item{\code{netherlands}}{a numeric vector}
    \item{\code{newzealand}}{a numeric vector}
    \item{\code{norway}}{a numeric vector}
    \item{\code{singapore}}{a numeric vector}
    \item{\code{spain}}{a numeric vector}
    \item{\code{sweden}}{a numeric vector}
    \item{\code{switzerland}}{a numeric vector}
    \item{\code{uk}}{a numeric vector}
    \item{\code{usa}}{a numeric vector}
  }
}

\examples{
data(conret)
}
\keyword{datasets}
