\name{drilling}
\Rdversion{1.1}
\alias{drilling}
\docType{data}
\title{
DRILLING dataset
}
\description{
Data on cost and attributes of offshore oil wells.
}
\usage{data(drilling)}
\format{
  A data frame with 26 observations on the following 7 variables.
  \describe{
    \item{\code{cost}}{cost ( $1,000)}
    \item{\code{depth}}{depth of well in ft}
    \item{\code{dril_eff}}{drilling efficiency measure}
    \item{\code{mud_wgt}}{mud wgt}
    \item{\code{angle}}{angle of drilling}
    \item{\code{tons_cas}}{tons of steel casing required}
    \item{\code{water_dp}}{water depth}
  }
}
\source{
MBA student in R. McCulloch's 322 class who worked for XXX Oil and Gas
}
\examples{
data(drilling)
}
\keyword{datasets}
