\name{drunk}
\alias{drunk}
\docType{data}
\title{
DRUNK dataset
}
\description{
Data collected by Prof. R. McCulloch through a questionaire
administered to two classes of 322.
}
\usage{data(drunk)}
\format{
  A data frame with 74 observations on the following 6 variables.
  \describe{
    \item{\code{nbeers}}{how many beers does it take you to get drunk?}
    \item{\code{wgt}}{weight of student}
    \item{\code{hgt}}{height of student}
    \item{\code{age}}{age of student}
    \item{\code{sex}}{sex of student}
    \item{\code{section}}{section of student}
  }
}
\examples{
data(drunk)
}
\keyword{datasets}
