\name{fiber_conn_sales}
\alias{fiber_conn_sales}
\docType{data}
\title{
FIBER_CONN_SALES dataset
}
\description{
Time series of sales of fiber optic connectors.
}
\usage{data(fiber_conn_sales)}
\format{
  A data frame with 265 observations on the following 2 variables.
  \describe{
    \item{\code{SALES}}{a numeric vector}
    \item{\code{time_trend}}{a numeric vector}
  }
}
\details{
a good example of heteroskedasticity around a linear time trend
}
\examples{
data(fiber_conn_sales)
}
\keyword{datasets}
