\name{hgtwgt}
\alias{hgtwgt}
\docType{data}
\title{
Height-Weight Data
}
\description{
Data on heights and weights of Kellogg MBAs
}
\usage{data(hgtwgt)}
\format{
  A data frame with 83 observations on the following 6 variables.
  \describe{
    \item{\code{sex}}{a factor with two levels "Male" and "Female"}
    \item{\code{swgt}}{student weight in pds}
    \item{\code{mhgt}}{mothers hgt (inches)}
    \item{\code{fhgt}}{fathers hgt (inches)}
    \item{\code{shgt}}{student hgt (inches)}
    \item{\code{phgt}}{mhgt+fhgt}
  }
}
\examples{
data(hgtwgt)
}
\keyword{datasets}
