\name{hotelocc}
\alias{hotelocc}
\docType{data}
\title{
HOTELOCC dataset
}
\description{
Variables are Chicago hotel occupancy rate and hotel X occupancy rate
}
\usage{data(hotelocc)}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{\code{Chi_occ}}{a numeric vector}
    \item{\code{HX_occ}}{a numeric vector}
  }
}
\examples{
data(hotelocc)
}
\keyword{datasets}
