\name{housepr}
\alias{housepr}
\docType{data}
\title{
HOUSEPR dataset
}
\description{
House prices and interest rates on mortgage for Chicago during the 12 months of 1986.
}
\usage{data(housepr)}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{Size}}{size of house in 1000 of sq ft}
    \item{\code{Price}}{price in \$1000}
  }
}
\examples{
data(housepr)
}
\keyword{datasets}
