\name{housepr_midcity}
\alias{housepr_midcity}
\docType{data}
\title{
HOUSEPR_MIDCITY dataset
}
\description{
Homes sold recently in MidCity from Dielman's regression text.
}
\usage{data(housepr_midcity)}
\format{
  A data frame with 128 observations on the following 29 variables.
  \describe{
    \item{\code{Home}}{a numeric vector}
    \item{\code{Nbhd}}{a numeric vector}
    \item{\code{Offers}}{a numeric vector}
    \item{\code{SqFt}}{a numeric vector}
    \item{\code{Brick}}{a factor with levels \code{No} \code{Yes}}
    \item{\code{Bedrooms}}{a numeric vector}
    \item{\code{Bathrooms}}{a numeric vector}
    \item{\code{Price}}{a numeric vector}
    \item{\code{pricethou}}{a numeric vector}
    \item{\code{sizethou}}{a numeric vector}
    \item{\code{Brickdum}}{a numeric vector}
    \item{\code{Nbhd_1}}{a numeric vector}
    \item{\code{Nbhd_2}}{a numeric vector}
    \item{\code{Nbhd_3}}{a numeric vector}
  }
}
\examples{
data(housepr_midcity)
}
\keyword{datasets}
