\name{lmich_yr}
\alias{lmich_yr}
\docType{data}
\title{
Lake Michigan Water Level Dataset
}
\description{
Annual water level in lake Michigan measured in June
}
\usage{data(lmich_yr)}
\format{
  A data frame with 89 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{a numeric vector}
    \item{\code{Level}}{Annual water level}
  }
}
\examples{
data(lmich_yr)
}
\keyword{datasets}
