\name{mammals}
\alias{mammals}
\docType{data}
\title{
MAMMALS dataset
}
\description{
The brain and body weights of 62 mammals.
}
\usage{data(mammals)}
\format{
  A data frame with 62 observations on the following 2 variables.
  \describe{
    \item{\code{bodywgt}}{body weight in kilograms}
    \item{\code{brainwgt}}{brain weight in grams}
  }
}
\source{
Data set 19 from Graphical Methods for Data Analysis by Chambers et al.
}
\examples{
data(mammals)
}
\keyword{datasets}
