\name{mileage}
\alias{mileage}
\docType{data}
\title{
Data on 2007 car attributes
}
\description{
Data on several product attributes for 2007 car models
}
\usage{data(mileage)}
\format{
  A data frame with 223 observations on the following 10 variables.
  \describe{
    \item{\code{Make}}{a factor with the make of the car}
    \item{\code{Model}}{a factor with the model of the car}
    \item{\code{Year}}{the year of the car}
    \item{\code{Curb.Weight..lbs.}}{the weight of the car}
    \item{\code{Horsepower}}{the horsepower of the car}
    \item{\code{Cylinder}}{the number of cylinders}
    \item{\code{Man.1..Auto.0.}}{1 if manual, 0 if automatic}
    \item{\code{Displacement..liters.}}{displacement/liters}
    \item{\code{EPA.MPG.City}}{EPA miles per gallon in the city}
    \item{\code{EPA.MPG.HW}}{EPA miles per gallon on highway}
  }
}
\source{
edmunds on-line
}
\examples{
data(mileage)
}
\keyword{datasets}
