\name{multi}
\alias{multi}
\docType{data}
\title{
MULTI dataset
}
\description{
Example of multiple regression 
}
\usage{data(multi)}
\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{p1}}{a numeric vector}
    \item{\code{p2}}{a numeric vector}
    \item{\code{Sales}}{a numeric vector}
  }
}
\details{
note the units of prices have been changed for confidentiality reasons.
}
\examples{
data(multi)
}
\keyword{datasets}
