\name{multicolinear}
\alias{multicolinear}
\docType{data}
\title{
MULTICOLINEAR dataset
}
\description{
Data with high multicolinearity
}
\usage{data(multicolinear)}
\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
  }
}
\examples{
data(multicolinear)
}
\keyword{datasets}
