\name{outliers}
\Rdversion{1.1}
\alias{outliers}
\docType{data}
\title{
OUTLIERS dataset
}
\description{
Data for linear regression with possible outliers and points of high leverage.
}
\usage{data(outliers)}
\format{
  A data frame with 104 observations on the following 4 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\examples{
data(outliers)
}
\keyword{datasets}
