\name{rat}
\alias{rat}
\docType{data}
\title{
RAT dataset
}
\description{
Data from the response of 19 rats to three are independent variables.
}
\usage{data(rat)}
\format{
  A data frame with 19 observations on the following 4 variables.
  \describe{
    \item{\code{x1}}{Indenpendent variable 1}
    \item{\code{x2}}{Indenpendent variable 2}
    \item{\code{x3}}{Indenpendent variable 3}
    \item{\code{Y}}{Rat response}
  }
}
\source{
Applied Linear Regression by Weisberg
}
\examples{
data(rat)
}
\keyword{datasets}
