\name{shock}
\alias{shock}
\docType{data}
\title{
SHOCK dataset
}
\description{
Shock absorber data from a major supplier of Ford.  The data is measurement
on the force transferred through a shock absorber.  Rebound measurements are
when the shank is forced out of the cylinder.  The measurements are taken
before and after the shock is filled with gas.

}
\usage{data(shock)}
\format{
  A data frame with 35 observations on the following 2 variables.
  \describe{
    \item{\code{reboundb}}{rebound before gas}
    \item{\code{rebounda}}{rebound after gas}
  }
}
\source{
George Easton
}
\examples{
data(shock)
}
\keyword{datasets}
