\name{triathlon}
\alias{triathlon}
\docType{data}
\title{
TRIATHLON dataset
}
\description{
Data from the 2003 Accenture Triathlon held on the Chicago Lakefront.
}
\usage{data(triathlon)}
\format{
  A data frame with 1107 observations on the following 7 variables.
  \describe{
    \item{\code{Last.Name}}{Participant's lastname}
    \item{\code{Sex}}{Participant's gender, either \code{F} or \code{M}}
    \item{\code{Class}}{Participant's class, with one or the following values: \code{F 12-14} \code{F 15-19} \code{F 20-24} \code{F 25-29} \code{F 30-34} \code{F 35-39} \code{F 40-44} \code{F 45-49} \code{F 50-54} \code{F 55-59} \code{F 60-64} \code{F 65-69} \code{M 12-14} \code{M 15-19} \code{M 20-24} \code{M 25-29} \code{M 30-34} \code{M 35-39} \code{M 40-44} \code{M 45-49} \code{M 50-54} \code{M 55-59} \code{M 60-64} \code{M 65-69} \code{M 70+}}
    \item{\code{Final}}{Total time.}
    \item{\code{Swim}}{Time for swimming stage.}
    \item{\code{Bike}}{Time for biking stage.}
    \item{\code{Run}}{Time for running stage.}
  }
}
\details{
look for Rossi 
}
\examples{
data(triathlon)
}
\keyword{datasets}
