\name{zagat_chi_rivernorth_08}
\Rdversion{1.1}
\alias{zagat_chi_rivernorth_08}
\docType{data}
\title{
ZAGAT_CHI_RIVERNORTH_08
}
\description{
Ratings on Chicago restaurants in the rivernorth area.
}
\usage{data(zagat_chi_rivernorth_08)}
\format{
  A data frame with 95 observations on the following 5 variables.
  \describe{
    \item{\code{name}}{the name of the restaurant, a factor variable}
    \item{\code{food}}{food rating}
    \item{\code{decor}}{decor rating}
    \item{\code{service}}{service rating}
    \item{\code{price}}{price rating}
  }
}
\examples{
data(zagat_chi_rivernorth_08)
}
\keyword{datasets}
